old_version/requirements.txt
langchain>=0.3.0
langchain-anthropic>=0.3.0
langgraph>=0.2.0
anthropic>=0.40.0
streamlit>=1.40.0
pydantic>=2.0.0
python-dotenv>=1.0.0
